Task2/requirements.txt
flask>=1.1.1
pytest>=5.3.5
requests>=2.23.0
healthcheck>=1.1
mysql-connector-python>=8.0.19
